docs: Add some missing docs to `Seeding2`
